Control.Scrobbler.Algorithm

Description

Scrobbling algorithm routines

Contesters

contest :: MonadIO m => Scrobbler m (PlayerStateChange Track) (Scrobble (Timed Track))

Check if candidate is ready to be scrobbled with default rules

contestWith :: MonadIO m => Rules -> Scrobbler m (PlayerStateChange Track) (Scrobble (Timed Track))

Check if candidate is ready to be scrobbled with custom rules

Scrobbling rules

type Rules = Int64 -> Timed Track -> Either ScrobblerError (Scrobble (Timed Track))

Scrobbling rules

First argument is time elapsed since scrobbler started

Second argument is contested candidate track

Result is either some scrobbler failure (most probably NoScrobble) or scrobbled track

Note: users should update _local themselves

Auxilliary

time' :: Monad m => Wire e m a Int64

Rounded time. Somehow useful